feat: return mapped JSON
